-- Created a LocalElemFE class to handle FE object for elements where 1D/2D elements can exist in 3D spaces. This is typically needed for structural and conduction elements. Earlier, this used to exist inside structural and conduction elements. -- LocalElemFE reimplements get_xyz() and get_normals() methods to return the locations in the global space, not the local space. This way, the elements don't need to worry about transformation. Currently, this has broken the interface for prestress function, which will need to be revisited. Careful consideration is also needed for development of composite material interface. -- Modified codes to be consistent with these modifications.

+++ src/aeroelasticity/pk_flutter_solver.cpp
@@ -269,7 +269,7 @@ MAST::PKFlutterSolver::print_sorted_roots()
         << "V      = " << std::setw(15) << std::setprecision(15) << root.V << std::endl
         << "g      = " << std::setw(15) << root.g << std::endl
         << "omega  = " << std::setw(15) << root.omega << std::endl
-        << std::setprecision(prec) // set the precision to the default value
+        << std::setprecision((int) prec) // set the precision to the default value
         << "Modal Participation : " << std::endl ;
         for (unsigned int j=0; j<nvals; j++)
             *_output
